requirements.txt
absl-py @ file:///opt/conda/conda-bld/absl-py_1639803114343/work
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
bleach==4.1.0
brotlipy==0.7.0
certifi==2021.10.8
cffi==1.15.0
cftime==1.6.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click @ file:///tmp/build/80754af9/click_1646056590078/work
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
cryptography @ file:///tmp/build/80754af9/cryptography_1639414572950/work
cx-Freeze==6.11.1
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==0.8.3
exoplasim==3.0.6
grpcio @ file:///tmp/build/80754af9/grpcio_1637590821884/work
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.12.0
ipykernel==6.9.1
ipython==8.1.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonpickle==2.0.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.0
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
jupyterthemes==0.20.0
kiwisolver==1.3.2
lesscpy==0.15.0
llvmlite==0.38.0
Markdown @ file:///tmp/build/80754af9/markdown_1614363852612/work
marko==1.2.1
markovify==0.9.3
MarkupSafe==2.1.0
matplotlib==3.4.3
matplotlib-inline==0.1.3
mido==1.2.10
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
more-itertools==8.10.0
music21==7.1.0
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
netCDF4==1.5.8
networkx==2.7.1
nltk @ file:///opt/conda/conda-bld/nltk_1645628263994/work
notebook==6.4.8
numba==0.55.1
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1634095651905/work
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
patchelf==0.14.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
ply==3.11
pretty-midi==0.2.9
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pynndescent==0.5.6
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.4
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
python-rtmidi==1.4.9
pytz==2022.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
regex @ file:///opt/conda/conda-bld/regex_1642021319040/work
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.0
similaritymeasures==0.4.4
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
stack-data==0.2.0
tensorboard==2.0.0
tensorboardX @ file:///tmp/build/80754af9/tensorboardx_1621440489103/work
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.1.0
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets==5.1.1
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
umap-learn==0.5.3
Unidecode==1.3.2
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wcwidth==0.2.5
webcolors==1.11.1
webencodings==0.5.1
Werkzeug @ file:///opt/conda/conda-bld/werkzeug_1645628268370/work
widgetsnbextension==3.5.2
zipp==3.8.0
